+++ src/TransmissonNET.App.Avalonia/Views/FileProgressBar.axaml.cs
@@ -1,0 +1,8 @@
+using Avalonia.Controls;
+
+namespace TransmissonNET.App.Avalonia.Views;
+
+public partial class FileProgressBar : UserControl
+{
+    public FileProgressBar() => InitializeComponent();
+}